refactor: extract remaining Args
